+++ Assets/UnityVoiceProcessor.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 24cef648ab228134486fc42b6ffaaf68
-NativeFormatImporter:
+guid: 0979be8284022dd40b41436688588942
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
-  mainObjectFileID: 0
   userData: 
   assetBundleName: 
   assetBundleVariant:
