+++ Assets/UnityVoiceProcessor/Demo.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: d459e6ce5df0548468d8be4bb381afa1
-TextScriptImporter:
+guid: 71bf916d382a2e24ea8365582186c390
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
