+++ Assets/UnityVoiceProcessor/Demo/README.md
@@ -1,0 +1,17 @@
+# Unity Voice Processor Demo
+
+This is an demo app that demonstrates how to ask for user permissions and capture output from `Unity Voice Processor`.
+
+## Compatibility
+
+- Android 5.0+ (API 21+) (ARM only)
+- iOS 11.0+
+- Windows (x86_64)
+- macOS (x86_64)
+- Linux (x86_64)
+
+## Usage
+
+The easiest way to run the demo is to simply import the [Unity Voice Processor Package](../../unity-voice-processor-1.0.0.unitypackage) into your project, open the `VoiceProcessorDemo` scene and hit play. To run on other platforms or in the player, go to _File > Build Settings_, choose your platform and hit the `Build and Run` button.
+
+Once the demo launches, press `Space` to start capturing audio. Press the number key between 0 and 3 to select a different device.
