+++ Assets/UnityVoiceProcessor/Demo/README.md.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: fb28db56de946b247a0fbaf414808aef
-DefaultImporter:
+guid: d35426ca9483b2e43ae0e9b1b9280949
+TextScriptImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
