+++ Assets/UnityVoiceProcessor/Demo/VoiceProcessorDemo.cs.meta
@@ -1,11 +1,9 @@
 fileFormatVersion: 2
-guid: 97db58b9403c0964c91782c4e42f3595
+guid: ea58e8594467a8f48b5248e94013c76a
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
-  defaultReferences:
-  - unityVoiceProcessorMixer: {fileID: 24300002, guid: 80f14401e61be2348b045082a2eec30d,
-      type: 2}
+  defaultReferences: []
   executionOrder: 0
   icon: {instanceID: 0}
   userData:
