+++ Assets/UnityVoiceProcessor/Demo/VoiceProcessorDemo.unity.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: d459e6ce5df0548468d8be4bb381afa1
-TextScriptImporter:
+guid: d364675f057970546810c3924beb9cd0
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
