+++ Assets/UnityVoiceProcessor/LICENSE.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: d459e6ce5df0548468d8be4bb381afa1
-TextScriptImporter:
+guid: fb28db56de946b247a0fbaf414808aef
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
