+++ Assets/UnityVoiceProcessor/MutePlaybackMixer.mixer.meta
@@ -1,7 +1,8 @@
 fileFormatVersion: 2
-guid: d459e6ce5df0548468d8be4bb381afa1
-TextScriptImporter:
+guid: 24cef648ab228134486fc42b6ffaaf68
+NativeFormatImporter:
   externalObjects: {}
+  mainObjectFileID: 0
   userData: 
   assetBundleName: 
   assetBundleVariant:
