+++ Assets/UnityVoiceProcessor/VoiceProcessorException.cs
