--- Demo.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 71bf916d382a2e24ea8365582186c390
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
