--- Demo/VoiceProcessorDemo.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: d364675f057970546810c3924beb9cd0
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
