--- Demo/VoiceProcessorListener.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 618483dd0c77d834486c41c358ff9707
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
