--- MutePlaybackMixer.mixer
@@ -1,69 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!241 &24100000
-AudioMixerController:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_Name: MutePlaybackMixer
-  m_OutputGroup: {fileID: 0}
-  m_MasterGroup: {fileID: 24300002}
-  m_Snapshots:
-  - {fileID: 24500006}
-  m_StartSnapshot: {fileID: 24500006}
-  m_SuspendThreshold: -80
-  m_EnableSuspend: 1
-  m_UpdateMode: 0
-  m_ExposedParameters: []
-  m_AudioMixerGroupViews:
-  - guids:
-    - 47b36b1fd5270d64e848709e771cb3b2
-    name: View
-  m_CurrentViewIndex: 0
-  m_TargetSnapshot: {fileID: 24500006}
---- !u!243 &24300002
-AudioMixerGroupController:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_Name: Master
-  m_AudioMixer: {fileID: 24100000}
-  m_GroupID: 47b36b1fd5270d64e848709e771cb3b2
-  m_Children: []
-  m_Volume: 7fbf53887df77d048850abcf7e148a38
-  m_Pitch: 87e550a1ca76d7b4180dccc98c9e942b
-  m_Send: 00000000000000000000000000000000
-  m_Effects:
-  - {fileID: 24400004}
-  m_UserColorIndex: 0
-  m_Mute: 0
-  m_Solo: 0
-  m_BypassEffects: 0
---- !u!244 &24400004
-AudioMixerEffectController:
-  m_ObjectHideFlags: 3
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_Name: 
-  m_EffectID: b19f9c4a20abfeb4ea1735063657e72d
-  m_EffectName: Attenuation
-  m_MixLevel: 4d90efd0617fbab4d9068a01fb2e344c
-  m_Parameters: []
-  m_SendTarget: {fileID: 0}
-  m_EnableWetMix: 0
-  m_Bypass: 0
---- !u!245 &24500006
-AudioMixerSnapshotController:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_Name: Snapshot
-  m_AudioMixer: {fileID: 24100000}
-  m_SnapshotID: 79614838aee912d49bd4c3e279fc015b
-  m_FloatValues:
-    7fbf53887df77d048850abcf7e148a38: -80
-  m_TransitionOverrides: {}
